\name{GateMeClass_train}
\alias{GateMeClass_train}
\title{GateMeClass_train}
\usage{
hello()
}
\description{
It inputs an annotated cytometry dataset in the form of an expression matrix and outputs a
marker table with a pseudo gating strategy for each cell type.
}
\examples{
For usage examples refer to https://github.com/simo1c/GateMeClass
}

\name{GateMeClass_annotate}
\alias{GateMeClass_annotate}
\title{GateMeClass_annotate}
\usage{
hello()
}
\description{
It inputs a dataset in the form of an expression matrix and a marker table and returns a vector of cell annotations.
}
\examples{
For usage examples refer to https://github.com/simo1c/GateMeClass
}
