\name{GateMeClass_annotate}
\alias{GateMeClass_annotate}
\title{GateMeClass_annotate}
\usage{
hello()
}
\description{
It inputs a dataset in the form of an expression matrix and a marker table and returns a vector of cell annotations.
}
\examples{
For usage examples refer to https://github.com/simo1c/GateMeClass
}
