\name{GateMeClass_train}
\alias{GateMeClass_train}
\title{GateMeClass_train}
\usage{
hello()
}
\description{
It inputs an annotated cytometry dataset in the form of an expression matrix and outputs a
marker table with a pseudo gating strategy for each cell type.
}
\examples{
For usage examples refer to https://github.com/simo1c/GateMeClass
}
